feat: add servicehoursexportmodel
