+++ interapp-backend/api/routes/endpoints/exports/exports.ts
@@ -1,4 +1,4 @@
-import { ExportsModel } from '@models/exports';
+import { AttendanceExportsModel } from '@models/exports';
 import { z } from 'zod';
 import { validateRequiredFieldsV2, verifyJWT, verifyRequiredPermission } from '@routes/middleware';
 import { ExportsFields } from './validation';
@@ -17,7 +17,7 @@ exportsRouter.get(
   async (req, res) => {
     const query = req.query as unknown as z.infer<typeof ExportsFields>;
 
-    const exports = await ExportsModel.packXLSX(query.id, query.start_date, query.end_date);
+    const exports = await AttendanceExportsModel.packXLSX(query.id, query.start_date, query.end_date);
 
     res.setHeader('Content-Type', xlsxMime);
     res.setHeader('Content-Disposition', 'attachment; filename=output.xlsx');
