+++ interapp-backend/tests/constants.test.ts
@@ -1,4 +1,4 @@
-import { ServiceModel, AuthModel, AnnouncementModel, UserModel, ExportsModel } from '../api/models';
+import { ServiceModel, AuthModel, AnnouncementModel, UserModel, AttendanceExportsModel, ServiceHoursExportsModel } from '../api/models';
 import { expect, test, describe } from 'bun:test';
 import { recreateDB, recreateMinio, recreateRedis } from './utils';
 
@@ -13,7 +13,7 @@ type TestSuite = {
 };
 
 // get all models in an array
-const objs = [ServiceModel, AuthModel, AnnouncementModel, UserModel, ExportsModel] as const;
+const objs = [ServiceModel, AuthModel, AnnouncementModel, UserModel, AttendanceExportsModel, ServiceHoursExportsModel] as const;
 
 // map all models to an object with the name as key
 const models = objs.reduce(
