+++ interapp-backend/tests/unit/AttendanceExportsModel.test.ts
@@ -1,9 +1,9 @@
 import { runSuite, testSuites } from '../constants.test';
-import { ExportsModel } from '@models/.';
+import { AttendanceExportsModel } from '@models/.';
 import { User } from '@db/entities';
 import { describe, test, expect } from 'bun:test';
 
-const SUITE_NAME = 'ExportsModel';
+const SUITE_NAME = 'AttendanceExportsModel';
 const suite = testSuites[SUITE_NAME];
 
 console.log(suite);
